Вынести логику создания алерта в AlertPresenter
