+++ MovieQuiz/Presentation/AlertPresenter.swift
@@ -12,6 +12,13 @@ final class AlertPresenter {
     weak var delegate: AlertPresenterDelegate?
     
     func showAlert(with model: AlertModel) {
-        delegate?.showAlert(with: model)
+        let alert = UIAlertController(title: model.title, message: model.message, preferredStyle: .alert)
+        let action = UIAlertAction(title: model.buttonText, style: .default) { [weak self] _ in
+            guard let self = self else { return }
+            model.completion?()
+            self.delegate?.dismissAlert()
+        }
+        alert.addAction(action)
+        delegate?.presentAlert(alert)
     }
 }
